Refactor WeakEventManager and enable nullable types
Refactored `WeakEventManager` to use instance-based methods,
removing the static `StaticWeakEventManager` class and
reintroducing it as a wrapper for backward compatibility.
Enabled nullable reference types in the project. Updated
target framework order to prioritize `netstandard2.1`.

Removed `StaticWeakEventManager.cs` from compilation.
Modernized code with lambda expressions, conditional
compilation, and cleanup of dead code. Enhanced unit tests
to cover new functionality, including support for custom
event argument types. Updated `Sender` and `Observer`
classes to reflect these changes.
