+++ Source/Scotec.Ssc/Tests/Scotec.Events.WeakEvents.Test/WeakEventManagerTest.cs
@@ -1,27 +1,38 @@
-﻿using Xunit;
+﻿using System;
+using Xunit;
 
 namespace Scotec.Events.WeakEvents.Test
 {
+    public class MyEventArgs : EventArgs
+    {
+
+    }
+
+
     public class WeakEventManagerTest
     {
         [Fact]
         public void Test()
         {
             var weakEventManager = new WeakEventManager();
-
+            
             var sender = new Sender();
-            StaticWeakEventManager.AddWeakHandler<Sender, EventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
-            StaticWeakEventManager.AddWeakHandler<Sender, EventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
+            weakEventManager.AddWeakHandler(sender, nameof(Sender.MyEvent), TestOnMyEvent);
+            weakEventManager.AddWeakHandler<Sender, MyEventArgs>(sender, nameof(Sender.MyEvent2), TestOnMyEvent2);
             sender.RaiseEvent();
 
-            StaticWeakEventManager.RemoveWeakHandler<Sender, EventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
+            weakEventManager.RemoveWeakHandler<Sender, MyEventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
             sender.RaiseEvent();
-            StaticWeakEventManager.RemoveWeakHandler<Sender, EventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
+            weakEventManager.RemoveWeakHandler<Sender, EventArgs>(sender, nameof(Sender.MyEvent), TestOnMyEvent);
             sender.RaiseEvent();
 
         }
 
+
         private void TestOnMyEvent(Sender sender, EventArgs e)
+        {
+        }
+        private void TestOnMyEvent2(Sender sender, MyEventArgs e)
         {
         }
 
@@ -33,10 +44,13 @@ namespace Scotec.Events.WeakEvents.Test
     public class Sender()
     {
         public event EventHandler<EventArgs>? MyEvent;
+        public event EventHandler<MyEventArgs>? MyEvent2;
+        //public event MyEventHandler<Sender, EventArgs>? MyEvent2;
 
         public void RaiseEvent()
         {
             MyEvent?.Invoke(this, new EventArgs());
+            MyEvent2?.Invoke(this, new MyEventArgs());
         }
     }
 
@@ -45,7 +59,7 @@ namespace Scotec.Events.WeakEvents.Test
     {
         public Observer(Sender test)
         {
-            StaticWeakEventManager.AddWeakHandler<Sender, EventArgs>(test, nameof(Sender.MyEvent), OnMyEvent);
+            //StaticWeakEventManager.AddWeakHandler<Sender, EventArgs>(test, nameof(Sender.MyEvent), OnMyEvent);
         }
 
         private void OnMyEvent(Sender sender, EventArgs e)
